add: simple crud api for courier management

+++ MasnaSzama/src/main/java/com/project/MasnaSzama/Repository/UserAccountRepository.java
@@ -3,6 +3,6 @@ package com.project.MasnaSzama.Repository;
 import com.project.MasnaSzama.Model.User.UserAccount;
 import org.springframework.data.repository.CrudRepository;
 
-public interface UserAccountRepository extends CrudRepository<UserAccount, Integer> {
+public interface UserAccountRepository extends CrudRepository<UserAccount, Long> {
 
 }
